requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
astroid==2.4.2
attrs==19.3.0
autobahn==20.6.2
Automat==20.2.0
black==19.10b0
Brotli==1.0.7
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
debugpy==1.0.0b12
Flask==1.1.2
Flask-Compress==1.5.0
Flask-SQLAlchemy==2.4.4
hyperlink==19.0.0
idna==2.10
incremental==17.5.0
isort==4.3.21
itemadapter==0.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
logparser==0.8.2
lxml==4.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
parsel==1.6.0
pathspec==0.8.0
pem==20.1.0
pexpect==4.8.0
Protego==0.1.16
psutil==5.7.0
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pylint==2.5.3
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
queuelib==1.5.0
regex==2020.7.14
requests==2.24.0
schedule==0.6.0
Scrapy==2.2.0
scrapyd==1.2.1
scrapydweb==1.4.0
service-identity==18.1.0
six==1.15.0
SQLAlchemy==1.3.18
SQLAlchemy-Utils==0.36.8
tabulate==0.8.7
toml==0.10.1
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
tzlocal==2.1
urllib3==1.25.9
w3lib==1.22.0
Werkzeug==1.0.1
wrapt==1.12.1
zope.interface==5.1.0
